requirements.txt
# Deployment requirements for Python 3.10 compatibility
python-dotenv==1.1.1
openai==1.98.0
anthropic==0.60.0
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-experimental==0.3.4
langchain-openai==0.3.28
langgraph==0.6.3
langgraph-checkpoint-sqlite==2.0.11
langsmith==0.4.10
semantic-kernel==1.32.2
autogen-agentchat==0.7.1
autogen-ext==0.7.1
gradio==5.39.0
ipywidgets==8.1.7
playwright==1.54.0
plotly==6.2.0
polygon-api-client==1.15.3
psutil==7.0.0
pypdf==5.9.0
pypdf2==3.0.1
requests==2.32.4
sendgrid==6.12.4
setuptools==80.9.0
smithery==0.1.0
speedtest-cli==2.1.3
wikipedia==1.4.0
bs4==0.0.2
lxml==6.0.0
mcp==1.12.3
mcp-server-fetch==2025.1.17
httpx==0.28.1
# Use ipython version compatible with Python 3.10
ipython==8.37.0
jupyter==1.1.1
ipykernel==6.30.0
